pyproject.toml
[project]
name = "tentai-show-pp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "imageio>=2.37.2",
    "pillow>=12.0.0",
]
